+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal.logger 0.2.0.9003
 
+* New function `log_shiny_input_changes` based on `logger` implementation, but curated to `teal` needs. 
+It allows to track all shiny inputs changes in teal modules on `TRACE` level with appended namespace name.
+
 # teal.logger 0.2.0
 
 * New function `register_handlers` to register global handlers for logging messages, warnings and errors.